extra charges api correct all

+++ src/pages/Bus/ShortOrders/Actions/Add.tsx
@@ -440,7 +440,7 @@ const AddBusShort = () => {
             extraChargePanel: []
         }));
         if (selectedCustomer && selectedVehicleClass) {
-            await fetch(`https://encodehertz.xyz/api/Long/GetExtraCharges?customerCode=${selectedCustomer}&vehicleClass=${selectedVehicleClass}`, {
+            await fetch(`https://encodehertz.xyz/api/Short/GetExtraCharges?customerCode=${selectedCustomer}&vehicleClass=${selectedVehicleClass}`, {
                 headers: {
                     'Authorization': `Bearer ${token}`,
                     'Content-Type': 'application/json'
